+++ examples/ecs-new-vpc/variables.tf
@@ -1,19 +1,4 @@
 # common variables
-variable "alicloud_access_key" {
-  description = "The Alicloud Access Key ID to launch resources."
-  default     = ""
-}
-
-variable "alicloud_secret_key" {
-  description = "The Alicloud Access Secret Key to launch resources."
-  default     = ""
-}
-
-variable "region" {
-  description = "The region to launch resources."
-  default     = "cn-shanghai"
-}
-
 variable "availability_zone" {
   description = "The available zone to launch ecs instance and other resources."
   default     = ""